初步功能都实现 但登录还有问题： 每条controller之前要有登录验证 授权后要重新登录才生效。

+++ src/main/java/com/chegy/dao/RoleDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.chegy.model.Role;
 
-public interface RoleDao extends JpaRepository<Role,Long>{
-
+public interface RoleDao extends JpaRepository<Role,Integer>{
+	
 }
