+++ src/main/java/com/chegy/dao/UserDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.chegy.model.User;
 
-public interface UserDao extends JpaRepository<User,Long>{
+public interface UserDao extends JpaRepository<User,Integer>{
 
 	User findByUsername(String username);
